[UPDATE] view basic dashboard information & best selling products
